refactor: move evaluateValue and Reference to evaluate utils
